% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculaIdade.R
\name{CalculaIdade}
\alias{CalculaIdade}
\title{Calculo da Idade}
\usage{
CalculaIdade(dia,mes,ano,diaref,mesref,anoref)
}
\arguments{
\item{dia}{Vetor com os dias de nascimento das pessoas (formato: dd)}

\item{mes}{Vetor com os meses de nascimento das pessoas (formato: mm)}

\item{ano}{Vetor com os anos de nascimento das pessoas (formato: aaaa)}

\item{diaref}{Dia de referencia (formato: dd)}

\item{mesref}{Mes de referencia (formato: mm)}

\item{anoref}{Ano de referencia (formato: aaaa)}
}
\description{
Funcao para calcular a idade dos estudantes considerando uma data de deferencia
}
\examples{

CalculaIdade(dia = c(24,30), mes = c(03,03), ano = c(1986,1986), diaref=25,mesref=03,anoref=2016)
}
