\name{CesgTools-package}
\alias{CesgTools-package}
\alias{CesgTools}
\docType{package}
\title{
\packageTitle{CesgTools}
}
\description{
\packageDescription{CesgTools}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{CesgTools}
\packageIndices{CesgTools}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{CesgTools}

Maintainer: \packageMaintainer{CesgTools}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%% ~~ simple examples of the most important functions ~~
}
