% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ContaChar.R
\name{ContaChar}
\alias{ContaChar}
\title{Conta o numero de vezes em que um conjunto de caracteres aparece}
\usage{
ContaChar(resp,vecchar=c(" ","*","."))
}
\arguments{
\item{resp}{\code{string} ou vetor de \code{strings} que sera avaliado}

\item{vecchar}{Vetor com os elementos que devem ser contabilizados.}
}
\description{
Conta o numero de vezes que qualquer um dos caracteres de \code{vecchar} aparece em
  uma string de texto
}
\examples{
ContaChar(" ABCD*    ....ADEDD")
ContaChar(" ABCD*    ....ADEDD",vecchar=" ")
ContaChar(" ABCD*    ....ADEDD",vecchar="*")
ContaChar(" ABCD*    ....ADEDD",vecchar=".")

}
