% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CTT1.R
\name{CTT1}
\alias{CTT1}
\title{Calcula as Estatisticas Classicas}
\usage{
CTT1(
  data,
  ids,
  vars,
  peso = FALSE,
  nitems,
  nforms,
  nitemform,
  nblform,
  nbl,
  tbl,
  gab,
  resp_possible,
  items,
  ndec = 2,
  acer = "1",
  erro = "0",
  napres = "9",
  mostra_napres = 0,
  calc_normit = FALSE
)
}
\arguments{
\item{data}{conjunto de dados a ser utilizado nas análises}

\item{ids}{vetor com os nomes das colunas de \code{data} que serao utilizadas
como identificacao na saida dos scores.}

\item{vars}{vetor com os nomes das colunas de \code{data} na ordem 
\code{c(caderno,respostas,peso (se houver))}.}

\item{nitems}{Numero de itens total.}

\item{nforms}{Numero de cadernos diferentes.}

\item{nitemform}{Numero de itens em cada caderno.}

\item{nblform}{Numero de blocos em que a prova estua organizada. Se eh uma prova
fixa, \code{nblform=1}.}

\item{nbl}{numero de blocos totais.7}

\item{tbl}{Tamanho do bloco, ou seja, numero de itens existentes no bloco.
Se eh uma prova fixa, \code{tbl=nitemform}.}

\item{gab}{Vetor com a string de gabaritos, cada item eh um elemento deste
vetor.}

\item{resp_possible}{Respostas possiveis (por exemplo:
\code{resp_possible = c(LETTERS[1:5],' ','*')}).}

\item{items}{Objeto do tipo matriz com \code{rownames} e
\code{colnames} definido.}

\item{ndec}{Numero de decimais para a saida das estatisticas classicas.
Default: \code{ndec=2}.}

\item{acer}{Codigo para a representacao do acerto no arquivo de saida.
Default: \code{acer='1'}}

\item{erro}{Codigo para a representacao do erro no arquivo de saida.
Default: \code{erro='0'}}

\item{napres}{Codigo para a representacao do item, quando nao apresentado
 ao estudante, no arquivo de saida. Default: \code{napres='9'}
@param peso vetor de pesos: Default: \code{peso = FALSE}
@param mostra_napres indicador sobre como os itens nao apresentados aos alunos
  devem ser exibidos. Default: \code{mostra_napres = 0}, ou seja o não apresentado
  é desconsiderado das médias dos alunos. \code{mostra_napres = 1} o não apresentado
  entra nas contas como se fosse uma resposta possível.#'}
}
\value{
O arquivo de saida desta funcao eh composto por 4 objetos
  dispostos em uma lista.

CTT[[1]] \code{data.frame} com as estatisticas classicas para
  todos os itens. Cada item eh uma linha deste objeto.

CTT[[2]] \code{data.frame} contendo: padrao de respostas:
  acerto = \code{acer}, erro = \code{erro} e
  item nao apresentado = \code{napres}; numero de acertos, numero de itens
  realizados pelo aluno, percentual de acertos e numero de acertos por bloco.

CTT[[3]] \code{data.frame} com as estatiticas classicas para os
  itens que podem ter apresentado algum problema. No caso, itens com 2 ou
  mais coeficientes bisseriais positivos e itens com coeficiente biserial
  abaixo de 0.15
}
\description{
Esta funcao calcula as Estatísticas Clássicas, score, score por bloco,
  percentual de acertos para um conjunto de dados e indica itens que possam
  ter algum tipo de problema como, por exemplo: gabarito incorreto, item
  mal elaborado, problema de impressao, entre outros.
}
\examples{
data("quest_dados")
data("gabpar05P")


item_prova <- matrix(c(1:36,13:36,1:12,25:36,1:24),byrow = T,ncol=36)
colnames(item_prova) <- paste('it',sprintf("\%02d",1:36),sep='')
rownames(item_prova) <- 1:nrow(item_prova)

dadosP <- quest_dados[nchar(quest_dados$rsp_por)==36,]
dadosP <- quest_dados[nchar(quest_dados$rsp_por)==36 & !is.na(quest_dados$cad_por),]

CTT1(
  data = dadosP,
  ids = c('codesc','turma','id'),
  vars = c('cad_por','rsp_por'),
  peso = FALSE,
  nitems = 36,
  nitemform = 36,
  nforms = 3,
  gab = gabpar05P$gab,
  resp_possible = c(LETTERS[1:4]," ","*"),
  items = item_prova,
  ndec = 2,
  nblform = 3,
  tbl = 12,
  nbl = 3,
  acer = '1',
  erro = '0',
  napres = '9',
  calc_normit = F)
   
}
\seealso{
\code{\link{RespItem}},\code{\link{Escore}},
         \code{\link{PontoBisserial}}, \code{\link{ItemPos}},
         \code{\link{write.ctt}}
}
