% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/not_in.R
\name{\%!in\%}
\alias{\%!in\%}
\title{Not in}
\usage{
x \%!in\% table
}
\arguments{
\item{x}{vetor}

\item{y}{vetor}
}
\value{
vetor de valores lógicos, se x está em y, etc.
}
\description{
Not in
}
\examples{
\dontrun{
x \%nin\% y}

}
