% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infoplogis3.R
\encoding{utf8}
\name{infoplogis3}
\alias{infoplogis3}
\title{Informacao do item sob o modelo logistico de 3 parametros}
\usage{
infoplogis3(theta, a, b, ci)
}
\arguments{
\item{theta}{Proficiência do indivíduo}

\item{a}{Discriminação do item}

\item{b}{Dificuldade do item}

\item{ci}{Parãmetro de acerto casual}
}
\value{
Informação do item
}
\description{
Informação do item utilizada nas demais funções do pacote
}
\details{
\code{Ii} Retorna a informação de Fisher do item com relação a proficiencia, para o modelo logístico de três parâmetros
}
