% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infotestplogis3.R
\encoding{utf8}
\name{infotestplogis3}
\alias{infotestplogis3}
\title{Informacao do teste para uma grade de proficiencias}
\usage{
infotestplogis3(a, b, ci, min.theta = -4, max.theta = 4)
}
\arguments{
\item{a}{Vetor de discriminações dos itens do teste}

\item{b}{vetor de dificuldades dos itens do teste}

\item{ci}{vetor de probabilidades de acerto casual dos itens do teste}

\item{min.theta}{inicio da grade (default = -4)}

\item{max.theta}{final da grade (default = 4)}
}
\value{
data.frame com duas colunas, theta com os valores de referencia da grade
e Info com as informações
}
\description{
Informação do teste, utilizada internamente na ATA
}
\details{
\code{\link{Iteste}} Retorna uma matriz de niveis de proficiencia e informacoes segundo o modelo
logistico de tres parametros
}
