% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normit.R
\name{normit}
\alias{normit}
\title{Calula o Normit}
\usage{
normit(scores, caderno, nitemform, peso)
}
\arguments{
\item{scores}{Vetor com os scores para o calculo do normit}

\item{caderno}{Vetor com os cadernos para cada um dos alunos (mesma ordem de scores)}

\item{nitemform}{numero de itens por caderno}

\item{peso}{peso para cada um dos alunos}
}
\value{
Retorna o normit
}
\description{
Esta funcao calcula o Normit para um dado conjunto
}
\examples{
normit...
}
