% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plogis3.R
\encoding{utf8}
\name{plogis3}
\alias{plogis3}
\title{Probabilidade de resposta sob o modelo logistico de 3 parametros}
\usage{
plogis3(theta, a, b, ci)
}
\arguments{
\item{theta}{Proficiência do indivíduo}

\item{a}{Discriminação do item}

\item{b}{Dificuldade do item}

\item{ci}{Parãmetro de acerto casual}
}
\value{
Probabilidade de acerto
}
\description{
Probabilidade de acerto sob o M3PL, utilizada nas demais funções desse pacote
}
\details{
\code{Pij} Retorna a probabilidade de um item dicotômico correto em função da proficiencia do indivíduo,
discriminação, dificuldade e probabilidade de acerto casual do item
}
