% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QuestGeral.R
\name{QuestGeral}
\alias{QuestGeral}
\title{Percentual de respondentes e media de desempenho para um conjunto de questoes}
\usage{
QuestGeral(
  base,
  prefixprofi,
  disc,
  vectitulo,
  vecquest,
  vecncat,
  veccat,
  quantil = c(0.25, 0.5, 0.75),
  pesopond = NULL,
  ndec = 2,
  bra = " ",
  inv = "*"
)
}
\arguments{
\item{base}{Conjunto de dados contendo o desempenho do estudante e as variaveis
a serem tabuladas no questionario.}

\item{prefixprofi}{Prefixo para as variaveis de proficiencia.}

\item{disc}{Vetor com o mnemonico para cada uma das disciplinas avaliadas.}

\item{vectitulo}{Vetor com as perguntas do questionario}

\item{vecquest}{Vetor com os nomes das variaveis (colunas) no conjunto de 
dados a ser tabulada.}

\item{vecncat}{Vetor com o numero das categorias das variaveis a serem tabuladas.}

\item{veccat}{Vetor com a descrição das categorias das variaveis a serem 
tabuladas.}

\item{quantil}{Quantis que serao computados, Default: \code{c(.25,;5,.75)}.}

\item{pesopond}{Vetor de pesos para expansao da amostra. Default: \code{NULL}}

\item{ndec}{Numero de casas decimais na saida da funcao.}

\item{branco}{Indicador do tipo de branco, Default: \code{" "}. Alternativamente,
pode-se utilizar \code{""}.}

\item{invalido}{Indicador do tipo de invalido, Default: \code{"*"}.}
}
\value{
Uma lista com objetos do tipo data.frame contendo as tabulacoes
 dos dados.
}
\description{
Retorna uma lista com cada elemento do tipo \code{data.frame} com o percentual
de respostas e media de desempenho para um conjunto de questoes.
}
\examples{
data("quest_dados")
data("quest_perguntas")
data("quest_categorias")
data("quest_numcategorias")

tabula <- QuestGeral(base = quest_dados,
           prefixprofi = 'profi_',
           disc = c('por','mat','nat','hum'),
           vectitulo = quest_perguntas ,
           vecquest = paste('p',sprintf("\%02d",1:20),sep=''),
           vecncat = quest_numcategorias,
           veccat = quest_categorias,
           quantil=c(.25,.5,.75),
           pesopond=NULL,
           ndec=2,
           bra=' ',
           inv='*')

tabula
}
