% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QuestProfi.R
\name{QuestProfi}
\alias{QuestProfi}
\title{Percentual de respondentes e media de desempenho para uma questao}
\usage{
QuestProfi(
  base,
  disc,
  quest,
  ncat,
  prefixprofi,
  quantil = c(0.25, 0.5, 0.75),
  pesopond,
  ndec,
  branco = " ",
  invalido = "*"
)
}
\arguments{
\item{base}{Conjunto de dados contendo o desempenho do estudante e as variaveis
a serem tabuladas no questionario.}

\item{disc}{Vetor com o mnemonico para cada uma das disciplinas avaliadas.}

\item{quest}{Nome da variaval (coluna) no conjunto de dados a ser tabulda.}

\item{ncat}{Numero de categorias da variavel a ser tabulada.}

\item{prefixprofi}{Prefixo para as variaveis de proficiencia.}

\item{quantil}{Quantis que serao computados, Default: \code{c(.25,;5,.75)}.}

\item{pesopond}{Vetor de pesos para expansao da amostra.}

\item{ndec}{Numero de casas decimais na saida da funcao.}

\item{branco}{Indicador do tipo de branco, Default: \code{" "}. Alternativamente,
pode-se utilizar \code{""}.}

\item{invalido}{Indicador do tipo de invalido, Default: \code{"*"}.}
}
\value{
Um objeto do tipo data.frame contendo a tabulacao dos dados.
}
\description{
Retorna um \code{data.frame} com o percentual de respostas e media de 
 desempenho por categoria da questao tabulada.
}
\examples{
data("quest_dados")
data("quest_perguntas")
data("quest_categorias")
data("quest_numcategorias")

i=1
QuestProfi(base = quest_dados,
           disc = c('por','mat','nat','hum'),
           quest = paste('p',sprintf("\%02d",1:20),sep='')[i],
           ncat = quest_numcategorias[i],
           prefixprofi = 'profi_',
           quantil=c(.25,.5,.75),
           pesopond=NULL,
           ndec=2,
           branco=' ',
           invalido='*')

}
