% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.blm_commands.R
\name{read.blm_commands}
\alias{read.blm_commands}
\title{Conversão das sintaxes de comando do Bilog}
\usage{
read.blm_commands(x)
}
\arguments{
\item{x}{vetor com a string do bilog a ser processada}
}
\value{
A saída desta função é uma lista nomeada com os parâmetros do BilogMG.
}
\description{
Função para, a partir de um vetor de texto de um comando de input do BilogMG,
  como: >GLOBAL, >SAVE, etc retorna os argumentos dos comandos em uma lista 
  nomeada.
}
\examples{
x <- c(">GLOBAL  DFNAME='E19_CH_V0.txt',","         NPARM = 3,","         SAVE;")

read.blm_commands(x)

}
\seealso{
\code{\link{read.blm_input}}
}
