% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.blm_input.R
\name{read.blm_input}
\alias{read.blm_input}
\title{Leitura da sintaxes do BIlog}
\usage{
read.blm_input(blm)
}
\arguments{
\item{blm}{caminho do arquivo .blm}
}
\value{
Objeto da classe blm_input contendo uma lista com os parâmetros de
 GLOBAL ,SAVE, LENGTH e INPUT.
}
\description{
Função para, a partir de um arquivo .blm retornar uma lista com os objetos 
  e parâmetros que estão sendo usados como input do Bilog.
}
