% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeBrancos.R
\name{removeBrancos}
\alias{removeBrancos}
\title{Resumo de medias}
\usage{
removeBrancos(x)
}
\arguments{
\item{x}{character ou vetor que deseja remover os espacos em branco}
}
\description{
Funcao que remove os brancos à esquerda ou à direita, equivalente ao trim...
}
\examples{
removeBrancos(" a  asda asd sss " )
removeBrancos(c(" a  asda asd sss "," asd2      344        ")

}
