% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resumoMedias.R
\name{resumoMedias}
\alias{resumoMedias}
\title{Resumo de medias}
\usage{
resumoMedias(dados,turmas,colmedia,colsresu,disc,probs = c(0,.1,.25,.5,.75,.9,1))
}
\arguments{
\item{dados}{\code{data.frame} contendo as informacoes a nivel dos alunos para 
o calculo das medias.}

\item{turmas}{\code{data.frame} contendo um cadastro com o menor nivel para juncao.
(Em dados escolares a turma eh o menor nivel de agregacao). Parametro nao obrigatorio.}

\item{colmedia}{Nome da coluna que contem a proficiencia em que a media serah calculada.}

\item{colsresu}{Vetor de caracteres com o nome das colunas das variaveis de \code{dados}
em ordem decrescente. Ex: anoescolar,uf,municipio,escola,turma}

\item{disc}{Mnemonico com a disciplina (util para identificacao na saida do arquivo)}

\item{probs}{Argumento com os quantis desejados para a saida da funcao.}
}
\description{
Funcao que calcula as medias por determinadas covariaveis hierarquicamente dispostas.
 Esta funcao calcula por exemplo eh capaz de calcular as medias totais, por escolas
 e turmas.
}
